backend/pyproject.toml
[project]
name = "secure-tasker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-axes>=8.1.0",
    "django-cors-headers>=4.9.0",
    "django-encrypted-model-fields>=0.6.5",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.3.2",
    "pytest-django>=4.11.1",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]
